+++ konva.js
@@ -8,7 +8,7 @@
    * Konva JavaScript Framework v4.0.6
    * http://konvajs.org/
    * Licensed under the MIT
-   * Date: Sat Aug 31 2019
+   * Date: Mon Sep 02 2019
    *
    * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
    * Modified work Copyright (C) 2014 - present by Anton Lavrenov (Konva)
@@ -6252,7 +6252,7 @@
               clearTimeout(this.dblTimeout);
               // Konva.inDblClickWindow = false;
           }
-          else {
+          else if (!DD.justDragged) {
               Konva.inDblClickWindow = true;
               clearTimeout(this.dblTimeout);
           }