+++ test/functional/TouchEvents-test.js
@@ -561,4 +561,66 @@ suite('TouchEvents', function() {
     assert.equal(tap, 2, 'tap triggered');
     assert.equal(dbltap, 1, 'no dbltap triggered');
   });
+
+  test('drag and second tap should not trigger dbltap', function() {
+    var stage = addStage();
+    var layer = new Konva.Layer();
+    stage.add(layer);
+
+    var circle1 = new Konva.Circle({
+      x: 100,
+      y: 100,
+      radius: 70,
+      fill: 'green',
+      stroke: 'black',
+      strokeWidth: 4,
+      name: 'myCircle1',
+      draggable: true
+    });
+    layer.add(circle1);
+    layer.draw();
+
+    var tap = 0;
+    var dbltap = 0;
+    var dragmove = 0;
+
+    stage.on('tap', e => {
+      assert.equal(e.target, circle1);
+      tap += 1;
+    });
+
+    stage.on('dbltap', e => {
+      dbltap += 1;
+    });
+
+    stage.on('dragmove', e => {
+      dragmove += 1;
+    });
+
+    stage.simulateTouchStart(
+      [{ x: 100, y: 100, id: 0 }],
+      [{ x: 100, y: 100, id: 0 }]
+    );
+
+    stage.simulateTouchMove(
+      [{ x: 150, y: 150, id: 0 }],
+      [{ x: 150, y: 150, id: 0 }]
+    );
+
+    stage.simulateTouchEnd([], [{ x: 150, y: 150, id: 0 }]);
+
+    assert.equal(tap, 0, 'no tap triggered');
+    assert.equal(dbltap, 0, 'no dbltap triggered');
+    assert.equal(dragmove, 1, 'dragmove triggered');
+
+    stage.simulateTouchStart(
+      [{ x: 150, y: 150, id: 0 }],
+      [{ x: 150, y: 150, id: 0 }]
+    );
+
+    stage.simulateTouchEnd([], [{ x: 150, y: 150, id: 0 }]);
+
+    assert.equal(tap, 1, 'tap triggered');
+    assert.equal(dbltap, 0, 'no dbltap triggered');
+  });
 });