+++ konva.js
@@ -8,7 +8,7 @@
    * Konva JavaScript Framework v4.0.6
    * http://konvajs.org/
    * Licensed under the MIT
-   * Date: Mon Sep 02 2019
+   * Date: Tue Sep 03 2019
    *
    * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
    * Modified work Copyright (C) 2014 - present by Anton Lavrenov (Konva)
@@ -2390,7 +2390,7 @@
                   if (distance < dragDistance) {
                       return;
                   }
-                  node.startDrag(evt);
+                  node.startDrag({ evt: evt });
                   // a user can stop dragging inside `dragstart`
                   if (!node.isDragging()) {
                       return;